Refactor drawing functions into separate files.
- Moved background drawing functions to `drawing/backgrounds.kt`.
- Moved pen stroke drawing functions to `drawing/penStrokes.kt`.
- Moved image drawing functions to `drawing/pageDrawing.kt`.
- Moved OpenGL rendering classes to `drawing/OpenGLRenderer.kt` and `drawing/LineRenderer.kt`.
